refactor: move close contact list to page component
