+++ pages/dashboard/tracing/[caseId]/index.tsx
@@ -1,8 +1,8 @@
 import { ArrowCircleLeftIcon } from "@heroicons/react/outline";
 import Link from "next/link";
-import { useRouter } from "next/router";
 import { CaseInformationItem } from "~/components/cases/case-information-item";
-import { CloseContactList } from "~/components/close-contact/close-contact-list";
+import { AddCloseContactButton } from "~/components/close-contact/add-close-contact-button";
+import { CloseContactCard } from "~/components/close-contact/close-contact-card";
 import {
   DashboardPage,
   DashboardPageHeader,
@@ -11,9 +11,9 @@ import {
 import { SectionCard, SectionCardHeader } from "~/components/ui/card";
 
 export default function TracingCaseDetail() {
-  const router = useRouter();
-
-  console.log(router.query.caseId);
+  const handleAddCloseContact = () => {
+    console.log("clicked");
+  };
 
   return (
     <DashboardPage>
@@ -50,7 +50,34 @@ export default function TracingCaseDetail() {
               </div>
             </div>
           </SectionCard>
-          <CloseContactList />
+          <div className="space-y-6">
+            <div className="md:flex md:items-center md:justify-between">
+              <div className="flex-1 min-w-0">
+                <h2 className="text-xl font-semibold leading-7 text-gray-900 sm:text-2xl sm:truncate">
+                  Daftar Kontak Erat Pasien
+                </h2>
+              </div>
+              <div className="hidden md:flex md:ml-4">
+                {/* TODO: add floating version for mobile */}
+                <AddCloseContactButton onClick={handleAddCloseContact} />
+              </div>
+            </div>
+            <div>
+              <div className="overflow-hidden rounded-md shadow bg-white">
+                <ul className="divide-y divide-gray-200">
+                  <li>
+                    <CloseContactCard />
+                  </li>
+                  <li>
+                    <CloseContactCard />
+                  </li>
+                  <li>
+                    <CloseContactCard />
+                  </li>
+                </ul>
+              </div>
+            </div>
+          </div>
         </div>
       </DashboardPageContent>
     </DashboardPage>
